Tab Group refactoring 3 of 100; leave element "black" in the font name ...

+++ src/lib/WP42HLListener.cpp
@@ -42,7 +42,7 @@ void WP42HLListener::insertCharacter(const guint16 character)
 	m_textBuffer.append(character);
 }
 
-void WP42HLListener::insertTab(const guint8 tabType, const guint16 tabPosition)
+void WP42HLListener::insertTab(const guint8 tabType, const float tabPosition)
 {
 	_flushText();
 	m_listenerImpl->insertTab();
